pyproject.toml
# SPDX-FileCopyrightText: 2026 PythonWoods <dev@pythonwoods.dev>
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zenzic"
version = "0.4.0rc2"
description = "Engineering-grade linter and strict build wrapper for MkDocs documentation sites"
readme = "README.md"
requires-python = ">=3.11"
# PEP 639: SPDX expression string (replaces deprecated `license = { text = "..." }`)
license = "Apache-2.0"
license-files = ["LICENSE", "LICENSES/*.txt"]
authors = [{ name = "PythonWoods", email = "dev@pythonwoods.dev" }]
keywords = ["documentation", "mkdocs", "zensical", "linter", "ci", "strict", "links"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27",
]


[project.scripts]
zenzic = "zenzic.main:cli_main"

[project.entry-points."zenzic.adapters"]
mkdocs   = "zenzic.core.adapters:MkDocsAdapter"
zensical = "zenzic.core.adapters:ZensicalAdapter"
vanilla  = "zenzic.core.adapters:VanillaAdapter"



[project.urls]
Homepage = "https://zenzic.pythonwoods.dev/"
Documentation = "https://zenzic.pythonwoods.dev/docs/"
Repository = "https://github.com/PythonWoods/zenzic"
Changelog = "https://github.com/PythonWoods/zenzic/blob/main/CHANGELOG.md"
Issues = "https://github.com/PythonWoods/zenzic/issues"

# ─── Build ────────────────────────────────────────────────────────────────────

[tool.hatch.build.targets.wheel]
packages = ["src/zenzic"]

# ─── Development dependencies (PEP 735 — managed by uv) ──────────────────────

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "nox>=2024.4.15",
    "reuse>=5.0.2",
    "pip-audit>=2.7.0",
    "bump-my-version>=1.2.6",
    "mkdocs-material[imaging]>=9.0.0",
    "mkdocs-minify-plugin>=0.7.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-static-i18n>=1.3.1",
    "types-pyyaml>=6.0.12.20250915",
    "requests>=2.33.0",  # CVE-2026-25645 — transitive via mkdocs-material/pip-audit
]

# ─── Ruff ─────────────────────────────────────────────────────────────────────

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP", "A"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["zenzic"]
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── Mypy ─────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true



[[tool.mypy.overrides]]
# BasePlugin.__init_subclass__ is untyped in MkDocs — unavoidable at subclass definition.
module = ["zenzic.plugin"]
disallow_untyped_calls = false

# ─── Pytest ───────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers --strict-config --import-mode=importlib"
testpaths = ["tests"]
markers = [
    "slow: marks slow tests (deselect with -m 'not slow')",
    "integration: marks tests that require external services or processes",
]

# ─── Coverage ─────────────────────────────────────────────────────────────────

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]

# ─── Version bumping ───────────────────────────────────────────────────────────

[tool.bumpversion]
current_version = "0.4.0rc2"
commit = true
tag = true
tag_name = "v{new_version}"
message = "release: bump version {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/zenzic/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "mkdocs.yml"
search = 'project_version: "{current_version}"'
replace = 'project_version: "{new_version}"'
